refactor: OpenID Connect에 맞게 패키지 이름 변경
