+++ src/main/java/aegis/server/global/security/oidc/UserAuthInfo.java
@@ -1,4 +1,4 @@
-package aegis.server.global.security.oauth;
+package aegis.server.global.security.oidc;
 
 import aegis.server.domain.member.domain.Member;
 import aegis.server.domain.member.domain.Role;
